refactor ReduceSpec: move PersistentStoreTester to separate file
