+++ rholang/src/test/scala/coop/rchain/rholang/interpreter/ReduceSpec.scala
@@ -1,21 +1,17 @@
 package coop.rchain.rholang.interpreter
 
-import java.nio.file.Files
-
 import cats.implicits._
 import com.google.protobuf.ByteString
 import coop.rchain.catscontrib.TaskContrib._
 import coop.rchain.crypto.codec.Base16
 import coop.rchain.crypto.hash.Blake2b512Random
-import coop.rchain.metrics
-import coop.rchain.metrics.{Metrics, NoopSpan, Span}
+import coop.rchain.metrics.NoopSpan
 import coop.rchain.models.Connective.ConnectiveInstance._
 import coop.rchain.models.Expr.ExprInstance._
 import coop.rchain.models.TaggedContinuation.TaggedCont.ParBody
 import coop.rchain.models.Var.VarInstance._
 import coop.rchain.models._
 import coop.rchain.models.rholang.implicits._
-import coop.rchain.rholang.interpreter.Runtime.RhoISpace
 import coop.rchain.rholang.interpreter.accounting._
 import coop.rchain.rholang.interpreter.errors._
 import coop.rchain.rholang.interpreter.storage.implicits._
@@ -23,75 +19,18 @@ import coop.rchain.rholang.Resources._
 import coop.rchain.rspace._
 import coop.rchain.rspace.{RSpace, ReplayRSpace}
 import coop.rchain.rspace.history.Branch
+import coop.rchain.rspace._
 import coop.rchain.rspace.internal.{Datum, Row, WaitingContinuation}
-import coop.rchain.shared.Log
-import coop.rchain.shared.PathOps._
 import monix.eval.Task
 import monix.execution.Scheduler.Implicits.global
 import org.scalatest.{Assertion, FlatSpec, Matchers}
 
 import scala.collection.immutable.BitSet
-import scala.collection.{SortedSet, mutable}
 import scala.collection.mutable.HashMap
+import scala.collection.{SortedSet, mutable}
 import scala.concurrent.Await
 import scala.concurrent.duration._
 
-final case class TestFixture(space: RhoISpace[Task], reducer: ChargingReducer[Task])
-
-trait PersistentStoreTester {
-
-  def withTestSpace[R](
-      errorLog: ErrorLog[Task]
-  )(f: TestFixture => R): R = {
-    val dbDir                              = Files.createTempDirectory("rholang-interpreter-test-")
-    implicit val logF: Log[Task]           = new Log.NOPLog[Task]
-    implicit val metricsEff: Metrics[Task] = new metrics.Metrics.MetricsNOP[Task]
-    implicit val noopSpan: Span[Task]      = NoopSpan[Task]()
-
-    implicit val cost = CostAccounting.emptyCost[Task].unsafeRunSync
-
-    val space = (RSpace
-      .create[
-        Task,
-        Par,
-        BindPattern,
-        ListParWithRandom,
-        ListParWithRandom,
-        TaggedContinuation
-      ](dbDir, 1024L * 1024L * 1024L, Branch("test")))
-      .unsafeRunSync
-    implicit val errLog = errorLog
-    val reducer         = RholangOnlyDispatcher.create[Task, Task.Par](space)._2
-    reducer.setPhlo(Cost.UNSAFE_MAX).runSyncUnsafe(1.second)
-    try {
-      f(TestFixture(space, reducer))
-    } finally {
-      space.close()
-      dbDir.recursivelyDelete()
-    }
-  }
-
-  def fixture[R](f: (RhoISpace[Task], ChargingReducer[Task]) => Task[R])(
-      implicit errorLog: ErrorLog[Task]
-  ): R = {
-    implicit val logF: Log[Task]           = new Log.NOPLog[Task]
-    implicit val metricsEff: Metrics[Task] = new metrics.Metrics.MetricsNOP[Task]
-    implicit val noopSpan: Span[Task]      = NoopSpan[Task]()
-    mkRhoISpace[Task]("rholang-interpreter-test-")
-      .use { rspace =>
-        for {
-          cost <- CostAccounting.emptyCost[Task]
-          reducer = {
-            implicit val c = cost
-            RholangOnlyDispatcher.create[Task, Task.Par](rspace)._2
-          }
-          _   <- reducer.setPhlo(Cost.UNSAFE_MAX)
-          res <- f(rspace, reducer)
-        } yield res
-      }
-      .runSyncUnsafe(3.seconds)
-  }
-}
 
 class ReduceSpec extends FlatSpec with Matchers with PersistentStoreTester {
   implicit val rand: Blake2b512Random = Blake2b512Random(Array.empty[Byte])
